first commit
This is a stone-paper-scissor guess game based on python. If you want run this code, you have to install python 3. Follow the steps to run this game.

How to run this game
-------------------------------------
1. Download or clone the folder
2. open the folder
3. open terminal or cmd in this folder or you can copy and paste the path
4. type python game.py
5. Enjoy the game

In case you want to modify it, you are always welcome.
If you want to contact with me, you can mail me.   scorpion004ss@gmail.com